refactor(test, pkg): src/__TEST__ -> test/, delete some unnecessary tests
