+++ test/wxMock.js
